Cleaning up Frame

+++ Engine/Source/Graphics/Managers/CommandBufferManager.cpp
@@ -19,7 +19,7 @@ CommandBufferManager::CommandBufferManager()
 
 	// Because one of the drawing commands involves binding the right VkFramebuffer, we'll actually have to record a command buffer for every image in the swap chain
 	mPerFramebufferCommandBuffers.reserve(gpSwapchainManager->mFramebuffers.size());
-	for (int64_t i = 0; i < gpSwapchainManager->mFramebuffers.size(); ++i)
+	for (int64_t i = 0; i < static_cast<int64_t>(gpSwapchainManager->mFramebuffers.size()); ++i)
 	{
 		mPerFramebufferCommandBuffers.emplace_back(i);
 	}
